refactor: rename http_executor to executor; create a middleware dir for doc.rs
